auth/auth_api/requirements.txt
flask==2.2.2
gevent==21.12.0
uvicorn==0.18.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
flask_sqlalchemy==2.5.1
Flask-JWT-Extended==4.4.3
Flask-Security==3.0.0
email-validator==1.2.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
apiflask==1.1.2
redis==4.3.4
click==8.1.3
pydantic[dotenv]==1.9.2
Flask-Migrate==3.1.0
pytest==7.1.2
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-instrumentation-sqlalchemy==0.29b0
user-agents==2.2.0
Flask-Limiter==2.6.2
Authlib==1.0.1
requests==2.28.1
sentry-sdk==1.9.10
ecs-logging==2.0.0